Dodano/Poprawiono obsługę masek 3x3 i 5x5.

+++ trunk/TypyDanych/Mask3x3/Interface/IMask3x3.h
@@ -11,8 +11,8 @@ using namespace std;
 class IMask3x3
 {
 	public:
-			static bool setValue(TypeConfig *type, int x, int y, double value);
-			static double getValue(TypeConfig *type, int x, int y);
+			static bool setValue(TypeConfig *type, int x, int y, int value);
+			static int getValue(TypeConfig *type, int x, int y);
 			static TypeConfig* getNew();
 };
